Set Up Framework and Minor Functionality/Testing
I've installed all the major technologies required for this project excluding database and overlay technologies. I'll be implementing those next. As is the program is capable of hosting a website from the frontend in both a desktop app and webview. It has a toggle for the web server that syncs using webhooks. and it uses basic http requests to transfer data to and from the backend.