Organized the file structure and switched to electron
changes on this branch may not be permanent. We'll see as development continues. electron seems to have all the functionality we want for both desktop and overlay. The overhead can be mitigated.

Main changes to structure was pulling all node modules (except the svelte instance), package.json files, package-lock.json files, and tsconfig filesinto the top project directory. Untestested as of yet, will do so tomorrow probably.